Fix race condition triggered on burying requests:
The server request MsgBuffer is now buried on enqueue_response() and not
when the request handler returns. Burying on request handler return can
cause a race condition as follows:

o Background request handler calls enqueue_response. The TX batch is
full, so a response is sent.

o Client receives the response and sends the next request for the sslot.
The request is received by the server's foreground thread. (Here, an assert
failure triggers because we expect the previous request MsgBuffer to be
buried at this point.)

o The background request handler for the initial request returns, and we
erroneously bury the request MsgBuffer of the second request.